requirements.txt
python-telegram-bot==21.3
gspread==6.1.2
google-auth==2.29.0
httpx==0.27.0
PyMuPDF==1.24.3
